requirements.txt
black==24.10.0
isort==5.13.2
Django>=5.0.3
djangorestframework==3.15.2
psycopg2-binary>=2.9.7
channels==4.1.0
firebase-admin==6.5.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0
asgiref==3.8.1
daphne==4.1.2
python-dotenv==1.0.1
django-unfold==0.50.0
jsonschema<=4.23.0
django-leaflet>=0.32.0